r/tests: extracted raft fixture base to be test framework independent
Previously `raft_fixture` was created with gtest in mind and its
dependencies made it impossible to be used in boost tests or
benchmarks. Separated a `raft_fixture_base` that can be used in
benchmark and boost tests.
